Create simple PropagationManager class to propagate TLEs
Note that we require a URL input to the constructor that points to
an API endpoint for TLEs. In addition, access propagations in the
`propagations` field, which is a dictionary mapping satellite names
to their position / velocity. Also, update propagations using the
`updatePropagations()` method. Note that this is WIP and not
efficient at all - NO interpolation right now.